Handle multiple combos pressed and released
This allows you to press and hold a combo, then hold another combo, ...
up to MAX_HELD_COMBOS (5). Then when you release any key, it releases
the key corresponding to the combo that that key was part of.

This is useful for gaming, when you want to use two WASD keys to move
diagonally.